feat: Add bfloat16 support (#136)

+++ tests/scripts/test_data_store.py
@@ -39,7 +39,7 @@ from mithril.models import (
 
 @pytest.mark.skip(reason="Move this test to DataStore method tests.")
 def test_data_store_1():
-    backend = TorchBackend(precision=32)
+    backend = TorchBackend()
     model = Linear(dimension=1)
     pm = PhysicalModel(
         model=model,
@@ -68,7 +68,7 @@ def test_data_store_1():
 @pytest.mark.skip(reason="Move this test to DataStore method tests.")
 def test_data_store_1_numpy():
     """Tests add_static_data works as expected for Numpy backend."""
-    backend = NumpyBackend(precision=32)
+    backend = NumpyBackend()
     model = Linear(dimension=1)
     pm = PhysicalModel(
         model=model,
@@ -100,7 +100,7 @@ def test_data_store_1_numpy():
 
 def test_data_store_3():
     """Tests all private attributes of DataStore are correct after compilation."""
-    backend = TorchBackend(precision=32)
+    backend = TorchBackend()
     model = Linear(dimension=1)
     static_data = {
         "input": backend.array([[1.0, 2, 3]]),
@@ -125,7 +125,7 @@ def test_data_store_4():
     corresponding keys. In this test, all inputs other than "output","_Shape_1_output"
     and "" should be unused.
     """
-    backend = TorchBackend(precision=32)
+    backend = TorchBackend()
     model = Model()
     model += Linear()(input="input", weight="weight", bias="bias")
     model += Shape()
@@ -155,7 +155,7 @@ def test_data_store_5():
     converts it only to corresponding backend tensor. So all keys other
     that "output" would become unused.
     """
-    backend = TorchBackend(precision=32)
+    backend = TorchBackend()
     model = Model()
     model += Linear()(input="input", weight="weight", bias="bias")
     model += Shape()
@@ -170,7 +170,7 @@ def test_data_store_6_error():
     """Tests if expected Exception raised when providing a static key in
     compile, if the key is an unusued key.
     """
-    backend = TorchBackend(precision=32)
+    backend = TorchBackend()
     model = Model()
     model += Linear()(input="input", weight="weight", bias="bias")
     model += Shape()
@@ -193,7 +193,7 @@ def test_data_store_7():
     # TODO: This test is expects cached_data to be "input" and "output" but
     # after we fix corresponding flat_graph handlings, it will be changed
     # to expect only "output" as cached_data and "input" as unused_keys.
-    backend = TorchBackend(precision=32)
+    backend = TorchBackend()
     model = Buffer()
 
     value = backend.array([[1.0, 2, 3]])
@@ -210,7 +210,7 @@ def test_data_store_7():
 
 
 def test_data_store_8():
-    backend = TorchBackend(precision=32)
+    backend = TorchBackend()
     model = Model()
     model += Sigmoid()(input="input", output=IOKey(name="output1"))
     model += Sigmoid()(input="input", output=IOKey(name="output2"))
@@ -227,7 +227,7 @@ def test_data_store_8():
 
 def test_data_store_9():
     """Infer static keys from pruned buffer"""
-    backend = TorchBackend(precision=32)
+    backend = TorchBackend()
     model = Model()
     model += Buffer()(input="input")
     model += Sigmoid()(input="input", output=IOKey(name="output1"))
@@ -246,7 +246,7 @@ def test_data_store_9():
 
 def test_data_store_10():
     """Infer static keys from pruned buffer 2"""
-    backend = TorchBackend(precision=32)
+    backend = TorchBackend()
     model = Model()
     model += Buffer()(input="input", output=IOKey(name="output1", expose=True))
     model += Sigmoid()(input="input", output=IOKey(name="output2", expose=True))
@@ -264,7 +264,7 @@ def test_data_store_10():
 
 
 def test_data_store_11():
-    backend = TorchBackend(precision=32)
+    backend = TorchBackend()
     model = Model()
     model += Sigmoid()(input="input", output=IOKey(name="output1", expose=True))
     model += Sigmoid()(input="input", output=IOKey(name="output2", expose=True))
@@ -285,7 +285,7 @@ def test_data_store_11():
 
 def test_data_store_13():
     """partial infer test"""
-    backend = TorchBackend(precision=32)
+    backend = TorchBackend()
     model = Model()
     add = Add()
     add.set_types(left=Tensor, right=Tensor)
@@ -315,7 +315,7 @@ def test_data_store_13():
 
 def test_data_store_14():
     """Infer statics with shapes"""
-    backend = TorchBackend(precision=32)
+    backend = TorchBackend()
     model = Model()
     model += Buffer()(input="input1", output=IOKey(name="out1", expose=True))
     model += (s := Shape())(input="out1")
@@ -371,7 +371,7 @@ def test_data_store_14():
 
 def test_data_store_15():
     """Infer statics with shapes"""
-    backend = TorchBackend(precision=32)
+    backend = TorchBackend()
     model = Model()
     model += Buffer()(input="input1", output=IOKey(name="out1", expose=True))
     model += (s := Shape())(input="out1")
@@ -427,7 +427,7 @@ def test_data_store_15():
 
 def test_data_store_16():
     """Tests add_static_data works as expected for Numpy backend."""
-    backend = NumpyBackend(precision=32)
+    backend = NumpyBackend()
     model = Linear(dimension=1)
     pm = PhysicalModel(
         model=model,
@@ -456,7 +456,7 @@ def test_data_store_16():
 
 def test_data_store_17():
     """Check 'runtime_static_keys'"""
-    backend = NumpyBackend(precision=32)
+    backend = NumpyBackend()
     model = Model()
     add = Add()
     add.set_types(left=Tensor, right=Tensor)
@@ -485,7 +485,7 @@ def test_data_store_17():
 
 def test_data_store_18():
     """Test infer ignore should remove from Data store 'runtime_static_keys'"""
-    backend = TorchBackend(precision=32)
+    backend = TorchBackend()
     model = Model()
     add = Add()
     add.set_types(left=Tensor, right=Tensor)
@@ -517,7 +517,7 @@ def test_data_store_18():
 
 def test_data_store_19():
     """Test infer ignore should remove infered data from Data store"""
-    backend = TorchBackend(precision=32)
+    backend = TorchBackend()
     model = Model()
     model += (add := Add())(left="left", right="right")
     model += Sigmoid()(input=add.output, output=IOKey("output"))
@@ -547,7 +547,7 @@ def test_data_store_19():
 
 def test_data_store_20():
     """Test data store holds intermediate non-differentiables correctly."""
-    backend = TorchBackend(precision=32)
+    backend = TorchBackend()
     model = Model()
     model += (add := Add())(left="left", right="right")
     model += (shp := Shape())(input=add.left)
